+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+# 0.1.0a2
+
+### Added
+
+* TOC directive accepts min_level and max_level of headings
+
+### Fixed
+
+* Border was not applied to the exec block if code was not shown
+
 # 0.1.0a1
 
-First release based on the dash-mantine-components v0.13.0a1.
+First release based on the dash-mantine-components v0.13.0a1